requirements.txt
# Core dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
google-analytics-data>=0.17.0
google-auth>=2.23.0
woocommerce>=3.0.0
requests>=2.31.0
python-dateutil>=2.8.2

# Configuration management
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0              # Retry logic
APScheduler>=3.10.0          # Task scheduling

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# v1.2 Features
reportlab>=4.0.8            # PDF export
openpyxl>=3.1.2             # Excel export
kaleido>=0.2.1              # Plotly to image conversion

# v1.3 Database - PostgreSQL support
psycopg2-binary>=2.9.0      # PostgreSQL driver (optional, for cloud deployment)
